migrate ResolvableMentionable to an interface
